requirements.txt
Pillow==7.1.2
webp-converter==4.0.1